+++ api/src/main/java/org/allaymc/api/block/property/type/BaseBlockPropertyType.java
@@ -1,9 +1,7 @@
 package org.allaymc.api.block.property.type;
 
 import lombok.Getter;
-import lombok.Setter;
 import lombok.ToString;
-import org.allaymc.api.block.property.BlockPropertyProcessor;
 
 import java.util.List;
 import java.util.Objects;
@@ -21,9 +19,6 @@ public abstract sealed class BaseBlockPropertyType<DATATYPE> implements BlockPro
     protected final List<DATATYPE> validValues;
     protected final DATATYPE defaultValue;
     protected final byte bitSize;
-
-    @Setter
-    protected BlockPropertyProcessor processor;
 
     protected BaseBlockPropertyType(String name, List<DATATYPE> validValues, DATATYPE defaultValue, byte bitSize) {
         this.name = name;