+++ server/src/main/java/org/allaymc/server/block/component/BlockStairsBaseComponentImpl.java
@@ -1,16 +1,16 @@
 package org.allaymc.server.block.component;
 
 import org.allaymc.api.block.BlockBehavior;
-import org.allaymc.api.block.BlockPlaceHelper;
 import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.property.type.BlockPropertyTypes;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.block.type.BlockType;
 import org.allaymc.api.world.Dimension;
+import org.allaymc.server.block.BlockPlaceHelper;
 import org.joml.Vector3ic;
 
-import static org.allaymc.api.block.BlockPlaceHelper.EWSN_DIRECTION_4_MAPPER;
+import static org.allaymc.server.block.BlockPlaceHelper.EWSN_DIRECTION_4_MAPPER;
 
 /**
  * @author daoge_cmd