+++ server/src/main/java/org/allaymc/server/block/component/crops/BlockPotatoesBaseComponentImpl.java
@@ -1,12 +1,12 @@
 package org.allaymc.server.block.component.crops;
 
 import org.allaymc.api.block.BlockBehavior;
-import org.allaymc.api.block.FortuneDropHelper;
 import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.block.type.BlockType;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.type.ItemTypes;
+import org.allaymc.server.block.FortuneDropHelper;
 
 import java.util.Set;
 import java.util.concurrent.ThreadLocalRandom;