+++ server/src/main/java/org/allaymc/server/block/component/grass/BlockShortGrassBaseComponentImpl.java
@@ -1,7 +1,6 @@
 package org.allaymc.server.block.component.grass;
 
 import org.allaymc.api.block.BlockBehavior;
-import org.allaymc.api.block.FortuneDropHelper;
 import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.block.type.BlockType;
@@ -9,6 +8,7 @@ import org.allaymc.api.entity.Entity;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.interfaces.ItemAirStack;
 import org.allaymc.api.item.type.ItemTypes;
+import org.allaymc.server.block.FortuneDropHelper;
 import org.allaymc.server.block.component.BlockBaseComponentImpl;
 
 import java.util.Set;