+++ server/src/main/java/org/allaymc/server/block/component/ore/BlockOreBaseComponentImpl.java
@@ -2,12 +2,12 @@ package org.allaymc.server.block.component.ore;
 
 import com.google.common.base.Preconditions;
 import org.allaymc.api.block.BlockBehavior;
-import org.allaymc.api.block.FortuneDropHelper;
 import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.block.type.BlockType;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.data.ItemId;
+import org.allaymc.server.block.FortuneDropHelper;
 import org.allaymc.server.block.component.BlockBaseComponentImpl;
 
 import java.util.Set;