+++ server/src/main/java/org/allaymc/server/block/component/trapdoor/BlockTrapdoorBaseComponentImpl.java
@@ -1,7 +1,6 @@
 package org.allaymc.server.block.component.trapdoor;
 
 import org.allaymc.api.block.BlockBehavior;
-import org.allaymc.api.block.BlockPlaceHelper;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.block.type.BlockType;
@@ -9,12 +8,13 @@ import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.math.MathUtils;
 import org.allaymc.api.world.Dimension;
 import org.allaymc.api.world.Sound;
+import org.allaymc.server.block.BlockPlaceHelper;
 import org.allaymc.server.block.component.BlockBaseComponentImpl;
 import org.joml.Vector3ic;
 
-import static org.allaymc.api.block.BlockPlaceHelper.EWSN_DIRECTION_4_MAPPER;
 import static org.allaymc.api.block.property.type.BlockPropertyTypes.DIRECTION_4;
 import static org.allaymc.api.block.property.type.BlockPropertyTypes.OPEN_BIT;
+import static org.allaymc.server.block.BlockPlaceHelper.EWSN_DIRECTION_4_MAPPER;
 
 /**
  * @author harry-xi